pyproject.toml
[project]
name = "users-service-ta"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cryptography>=43.0.3",
    "databases>=0.9.0",
    "dependency-injector>=4.43.0",
    "fastapi[all]>=0.115.5",
    "opentelemetry-instrumentation-fastapi>=0.49b2",
    "opentelemetry-instrumentation-sqlalchemy>=0.49b2",
    "pymysql>=1.1.1",
    "pytest>=8.3.3",
    "sqlmodel>=0.0.22",
]
